requirements.txt
django==3.2.3
python-dotenv==0.21.0
